split states
