+++ feature3/fooddiary/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/fooddiary/ui/goals/settings/DailyGoalsFormFieldError.kt
@@ -1,0 +1,6 @@
+package com.maksimowiczm.foodyou.feature.fooddiary.ui.goals.settings
+
+internal enum class DailyGoalsFormFieldError {
+    Required,
+    NotANumber
+}
