+++ feature3/fooddiary/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/fooddiary/ui/goals/settings/DailyGoalsScreen.kt
@@ -1,6 +1,5 @@
 package com.maksimowiczm.foodyou.feature.fooddiary.ui.goals.settings
 
-import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -12,58 +11,38 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.outlined.Undo
-import androidx.compose.material.icons.outlined.Info
 import androidx.compose.material.icons.outlined.Percent
-import androidx.compose.material.icons.outlined.Warning
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ButtonGroupDefaults
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.ToggleButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.role
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.maksimowiczm.foodyou.core.preferences.collectAsStateWithLifecycleInitialBlock
 import com.maksimowiczm.foodyou.core.preferences.userPreference
 import com.maksimowiczm.foodyou.core.ui.ArrowBackIconButton
 import com.maksimowiczm.foodyou.core.ui.ext.add
-import com.maksimowiczm.foodyou.core.ui.form.FormField
-import com.maksimowiczm.foodyou.core.ui.theme.LocalNutrientsPalette
 import com.maksimowiczm.foodyou.core.util.NutrientsHelper
+import com.maksimowiczm.foodyou.feature.food.domain.NutritionFactsField
+import com.maksimowiczm.foodyou.feature.fooddiary.domain.DailyGoal
 import com.maksimowiczm.foodyou.feature.fooddiary.preferences.GoalsPreference
 import foodyou.app.generated.resources.*
 import kotlin.math.roundToInt
-import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.transformLatest
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 
@@ -92,7 +71,7 @@ internal fun DailyGoalsScreen(onBack: () -> Unit, modifier: Modifier = Modifier)
         ) {
             item {
                 DailyGoalsForm(
-                    formState = state,
+                    goal = weeklyGoals.monday,
                     contentPadding = PaddingValues(horizontal = 16.dp)
                 )
             }
@@ -103,10 +82,31 @@ internal fun DailyGoalsScreen(onBack: () -> Unit, modifier: Modifier = Modifier)
 @OptIn(ExperimentalCoroutinesApi::class, ExperimentalMaterial3ExpressiveApi::class)
 @Composable
 internal fun DailyGoalsForm(
-    formState: DailyGoalsFormState,
+    goal: DailyGoal,
     contentPadding: PaddingValues,
     modifier: Modifier = Modifier
 ) {
+    var useDistribution by rememberSaveable { mutableStateOf(false) }
+
+    val proteins = goal[NutritionFactsField.Proteins].toFloat()
+    val carbohydrates = goal[NutritionFactsField.Carbohydrates].toFloat()
+    val fats = goal[NutritionFactsField.Fats].toFloat()
+    val energy = goal[NutritionFactsField.Energy].roundToInt()
+
+    val sliderState = rememberMacroInputSliderFormState(
+        proteins = NutrientsHelper.proteinsPercentage(energy, proteins) * 100f,
+        carbohydrates = NutrientsHelper.carbohydratesPercentage(energy, carbohydrates) * 100f,
+        fats = NutrientsHelper.fatsPercentage(energy, fats) * 100f,
+        energy = energy
+    )
+
+    val weightState = rememberMacroWeightInputFormState(
+        proteins = proteins,
+        carbohydrates = carbohydrates,
+        fats = fats,
+        energy = energy
+    )
+
     Column(modifier) {
         Text(
             text = "Pick the days",
@@ -145,43 +145,22 @@ internal fun DailyGoalsForm(
             color = MaterialTheme.colorScheme.primary
         )
         DistributionButtons(
-            formState = formState,
+            useDistribution = useDistribution,
+            onUseDistributionChange = { useDistribution = it },
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(contentPadding)
         )
-        if (formState.useDistribution) {
-            val energy = NutrientsHelper.calculateEnergy(
-                proteins = formState.proteins.value,
-                carbohydrates = formState.carbohydrates.value,
-                fats = formState.fats.value
-            ).roundToInt()
-
-            val state = rememberMacroInputSliderFormState(
-                proteins = NutrientsHelper.proteinsPercentage(
-                    energy = energy,
-                    proteins = formState.proteins.value
-                ) * 100f,
-                carbohydrates = NutrientsHelper.carbohydratesPercentage(
-                    energy = energy,
-                    carbohydrates = formState.carbohydrates.value
-                ) * 100f,
-                fats = NutrientsHelper.fatsPercentage(
-                    energy = energy,
-                    fats = formState.fats.value
-                ) * 100f,
-                energy = energy
-            )
-
+        if (useDistribution) {
             MacroInputSliderForm(
-                state = state,
+                state = sliderState,
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(contentPadding)
             )
         } else {
-            WeightForm(
-                formState = formState,
+            MacroWeightInputForm(
+                state = weightState,
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(contentPadding)
@@ -192,7 +171,11 @@ internal fun DailyGoalsForm(
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
-private fun DistributionButtons(formState: DailyGoalsFormState, modifier: Modifier = Modifier) {
+private fun DistributionButtons(
+    useDistribution: Boolean,
+    onUseDistributionChange: (Boolean) -> Unit,
+    modifier: Modifier = Modifier
+) {
     Row(
         modifier = modifier,
         horizontalArrangement = Arrangement.spacedBy(
@@ -201,8 +184,8 @@ private fun DistributionButtons(formState: DailyGoalsFormState, modifier: Modifi
         )
     ) {
         ToggleButton(
-            checked = !formState.useDistribution,
-            onCheckedChange = { formState.useDistribution = !it },
+            checked = !useDistribution,
+            onCheckedChange = { onUseDistributionChange(!it) },
             modifier = Modifier
                 .height(56.dp)
                 .semantics { role = Role.RadioButton }
@@ -213,8 +196,8 @@ private fun DistributionButtons(formState: DailyGoalsFormState, modifier: Modifi
             )
         }
         ToggleButton(
-            checked = formState.useDistribution,
-            onCheckedChange = { formState.useDistribution = it },
+            checked = useDistribution,
+            onCheckedChange = { onUseDistributionChange(it) },
             modifier = Modifier
                 .height(56.dp)
                 .semantics { role = Role.RadioButton }
@@ -225,221 +208,4 @@ private fun DistributionButtons(formState: DailyGoalsFormState, modifier: Modifi
             )
         }
     }
-}
-
-@OptIn(ExperimentalCoroutinesApi::class)
-@Composable
-private fun WeightForm(formState: DailyGoalsFormState, modifier: Modifier = Modifier) {
-    Column(modifier) {
-        AnimatedVisibility(formState.autoCalculateEnergy) {
-            Column {
-                Spacer(Modifier.height(8.dp))
-                AutoCalculateCard(
-                    formState = formState,
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-        var badEnergy by rememberSaveable { mutableStateOf(false) }
-        LaunchedEffect(formState) {
-            snapshotFlow { formState.badEnergy }.transformLatest {
-                if (it) {
-                    delay(1.seconds)
-                }
-
-                emit(it)
-            }.collectLatest {
-                badEnergy = it
-            }
-        }
-        AnimatedVisibility(badEnergy) {
-            Column {
-                Spacer(Modifier.height(8.dp))
-                BadEnergyCard(
-                    formState = formState,
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-        Spacer(Modifier.height(8.dp))
-        MacroInput(formState)
-    }
-}
-
-@Composable
-private fun BadEnergyCard(formState: DailyGoalsFormState, modifier: Modifier = Modifier) {
-    Surface(
-        modifier = modifier,
-        color = MaterialTheme.colorScheme.errorContainer,
-        contentColor = MaterialTheme.colorScheme.onErrorContainer,
-        shape = MaterialTheme.shapes.medium
-    ) {
-        Column(
-            modifier = Modifier.padding(
-                top = 16.dp,
-                start = 16.dp,
-                end = 16.dp,
-                bottom = 8.dp
-            ),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Warning,
-                contentDescription = null
-            )
-            Text(
-                text = "The total calories from your proteins, carbohydrates, and fat don't add up to your daily energy goal",
-                style = MaterialTheme.typography.bodyMedium
-            )
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.End
-            ) {
-                Button(
-                    onClick = { formState.autoSetEnergy() },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.onErrorContainer,
-                        contentColor = MaterialTheme.colorScheme.errorContainer
-                    )
-                ) {
-                    Text("Fix it")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun AutoCalculateCard(formState: DailyGoalsFormState, modifier: Modifier = Modifier) {
-    Surface(
-        modifier = modifier,
-        color = MaterialTheme.colorScheme.surfaceContainer,
-        contentColor = MaterialTheme.colorScheme.onSurface,
-        shape = MaterialTheme.shapes.medium
-    ) {
-        Column(
-            modifier = Modifier.padding(
-                top = 16.dp,
-                start = 16.dp,
-                end = 16.dp,
-                bottom = 8.dp
-            ),
-            horizontalAlignment = Alignment.End
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    imageVector = Icons.Outlined.Info,
-                    contentDescription = null
-                )
-                Spacer(Modifier.width(8.dp))
-                Text(
-                    text = "Energy is calculated automatically based on macronutrients",
-                    style = MaterialTheme.typography.bodyMedium
-                )
-            }
-            TextButton(
-                onClick = {
-                    formState.autoCalculateEnergy = false
-                }
-            ) {
-                Text("I want to set energy manually")
-            }
-        }
-    }
-}
-
-@Composable
-private fun MacroInput(formState: DailyGoalsFormState, modifier: Modifier = Modifier) {
-    val nutrientsPalette = LocalNutrientsPalette.current
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        formState.proteins.TextField(
-            label = stringResource(Res.string.nutriment_proteins),
-            color = nutrientsPalette.proteinsOnSurfaceContainer,
-            modifier = Modifier.fillMaxWidth()
-        )
-        formState.carbohydrates.TextField(
-            label = stringResource(Res.string.nutriment_carbohydrates),
-            color = nutrientsPalette.carbohydratesOnSurfaceContainer,
-            modifier = Modifier.fillMaxWidth()
-        )
-        formState.fats.TextField(
-            label = stringResource(Res.string.nutriment_fats),
-            color = nutrientsPalette.fatsOnSurfaceContainer,
-            modifier = Modifier.fillMaxWidth()
-        )
-        if (formState.autoCalculateEnergy) {
-            val energy = NutrientsHelper.calculateEnergy(
-                proteins = formState.proteins.value,
-                carbohydrates = formState.carbohydrates.value,
-                fats = formState.fats.value
-            )
-
-            val str = buildString {
-                append("=")
-                append(" ${energy.roundToInt()} ")
-                append(stringResource(Res.string.unit_kcal))
-            }
-
-            Text(
-                text = str,
-                modifier = Modifier.fillMaxWidth(),
-                textAlign = TextAlign.Center,
-                style = MaterialTheme.typography.headlineSmall
-            )
-        } else {
-            OutlinedTextField(
-                state = formState.energy.textFieldState,
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(Res.string.unit_energy)) },
-                suffix = { Text(stringResource(Res.string.unit_kcal)) },
-                keyboardOptions = KeyboardOptions(
-                    keyboardType = KeyboardType.Decimal,
-                    imeAction = ImeAction.Next
-                ),
-                trailingIcon = {
-                    IconButton(
-                        onClick = { formState.autoCalculateEnergy = true }
-                    ) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Outlined.Undo,
-                            contentDescription = null
-                        )
-                    }
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun FormField<Float, DailyGoalsFormFieldError>.TextField(
-    label: String,
-    color: Color,
-    modifier: Modifier = Modifier,
-    suffix: String = stringResource(Res.string.unit_gram_short),
-    imeAction: ImeAction = ImeAction.Next
-) {
-    OutlinedTextField(
-        state = textFieldState,
-        modifier = modifier,
-        label = { Text(label) },
-        suffix = { Text(suffix) },
-        keyboardOptions = KeyboardOptions(
-            keyboardType = KeyboardType.Decimal,
-            imeAction = imeAction
-        ),
-        colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = color,
-            unfocusedBorderColor = color,
-            focusedLabelColor = color,
-            unfocusedLabelColor = color
-        )
-    )
 }
